Put some debug prints, fix Camera Pos update using deltas
Instead of adding I was multiplying the deltas to camera position,
this has been fixed now.

The camera.getpos, doesnt seem to be returning absolute position
values, but instead returning relative to its last position or so,
need to look at this bit more. Because of this, I trying to control
camera isnt working. So I am continuing with using the panda3d
builtin keyboard based control logic for now (i.e using useDrive
call to set that thing up)