Rename movement deltas to ctrans, also cpos
The camera represents the main characters perspective in the game,
So have renamed the movement translation deltas related var as
ctrans.

As camera or rather any node's simple getPos doesnt seem to get the
absolute position of that object in the 3d world (that one has set
using setPos), when processing/calling from the frame related task
callbacks, so now I maintain the expected camera absolute position
in a variable called cpos.

NOTE: I need to also maintain the camera/character's rotation
transform value. Also once rotation support is added, then movement
of camera/character needs to account of the rotation factor also
and not just a simple blind translation ie both translation and
rotation need to be accounted in order to determine the new position.