Start stopped, slower movement change control
Now the camera/main_character starts with no movement. User has to
use the keys to add movement as required.

Have reduced the speed added or removed wrt the translations.

Have added variables to store the camera facing direction related
info as well as deltas to use wrt changing orientation.

Calling getPos with self.render as the base doesnt seem to help
get the absolute position. Maybe have to have a explict dummy node
at 0,0,0 and then see if trying to getPos relative to it helps.