Rather disableMouse does the trick
Rather the builtin mouse based camera handler seems to be messing
with getting the camera position, maybe it handles things with
some additional indirection or so. However on disabling this,
the camera position seems to print properly.

However retaining the internal cpos, cface vars for now, as it can
allow a out of body 3rd person viewing if required later, where
main character (rather plane eventually) and camera can be at diff
locations.