Avoid maintaining CurPosAndOrientation wrt Camera/Character
Rather initially as I was not able to get the actual position of
the camera from camera itself, after it has undergone a series of
transforms, so I had maintained a seperate Vector ie a set of 3
values, which represent the current position of the camera after
it has undergone all the transforms.

Till now I was doing only the translation transform. And translation
inturn as a Vec3 can be applied directly to the Camera Vec3 to get
the resultant Camera Vec3.

Now I wanted to add rotation transform, but then it requires a Mat3
to be defined which contains the required rotation related transform
and inturn Matrix multiplication to be done on the Vec3 representing
the current position of Camera.

Mat3 class provides a rotateMat function to which one can give the
amount of rotation required and the axis around which rotation
should occur. And it will return the 3x3 matrix which represents
the rotation transform.

But then the Mat3 class of P3d doesnt overload * (ie _mul_ func in
python) to allow multiplying with a Vec3 directly.

[POST MODIFICATIONS: As the rotateMat's 2nd argument, I hadnt looked
at in the document and so was behaving bit strange, I looked into
the Mat3 class finally [[ RTFM ;( ]], and realised the meaning of the
2nd argument to rotateMat, as well as saw the xform function. So
bit lately realised that there is a xform function provided, which
one can call to multiple the Mat3 with the Vec3 i.e Mat3.xform(Vec3)]

Also to cumulate all the operations/transforms on a object in 3d space
till that point in time, one will require to use a Mat4 (the homogeneous
version), ie one which can hold the cumulative effects of translations,
rotations, scaling, till that point in time. So I tried Mat4, and that
also didnt allow multiplying with a Vec3 or even the more appropriate
Vec4 directly.

WIth the above, given that my original issue with getting position of
camera is no longer there, given that I am using disableMouse to avoid
P3Ds internal camera movement handler, and thus I have full freedom
wrt camera movement, as well as getting its current Position, after
applying transforms. So I have decided to directly use the internal
position, orientation info got from camera's internal transform matrix
itself to represent it rather than the duplicate cpos (and cface)
which I was using.

This commit represents the changes to get the cumulated cpos and the
direct translation transform which I was applying on it, being replaced
by applying the ctrans relative transform to camera directly (by using
camera itself as the node about which to apply the relative translation
transform).

Dont forget the POSTMODIFICATION Note above, which would have allowed
me to do Mat4.xform(Vec4) to achieve what I initially wanted, when
I started on the work related to this commit, and avoid some of the
overall semantic changes brought about by this commit.