SetFluidPos/SetPos, Larger generated terrain
As the motion was jerky with a 4Kx4K loaded terrain, so crosschecked
the same size with a generated terrain. Both seem equally slow.

Tried changing from setFluidPos to setPos, as the later is not
supposed to check for collisions, but that didnt change the performance.
This is as expected at one level, but was just tried, to be sure.