Fix a oversight with color shading logic
The TopView and BruteForceFullLOD added in last commit to help
see how things are, helped identify certain issues which are
fixed here.

The sea and below level I was expecting to be between 0 to 0.2
gray scale value in the heightfield, but the one which I had
adjusted for use currently has see level not reaching upto 0.2,
so reduced to 0.1 for now.

The way I was handling the green and red zones was not proper,
have fixed it to some extent now.

NOTE: However for things which are towards the lower band of the
green and red zones, may appear black, need to fix this aspect.