Set heightfield after colormap is generated
As I notice some issue with the genered terrain wrt color and height,
so wanted to check if the autoscaling of heightfield to be a power
of 2 +1, was creatIng some issue. ie the pixel positions in hf was
mapping to proper corresponding one in cm or not. So I loaded hf
into a PNMImage and then created corresponding cm and then set both
to the terrain logic.

But however my issue with color and height in the generated terrain
still remains. Need to check my colormap generation logic as well
as read about the terrain generator once.