Move creating heightfield and colormap into independent funcs
Also the filename for terrain no longer gives full heightfield
file name. Rather now it gives a base filename from which the
heightfield and colormap filenames can be derived and used.

With this commit, now user can also load a associated colormap
file for a given heightfield file, if available.