If there is no cm and its autogenerated, then save it.
NOTE: Also the heightfield and colormap seem to map to one another
by default, only when they are vertically flipped wrt one another.
So any autogenerated colormap will by default appear flipped wrt
the terrain heightfield, and one will have to flip the saved img
to get it to align with terrain in the program for now. Need to
look at changing the way the normals/... for the individual layers/
entities like heightfield, colormap etc are defined so that both
will be aligned.