Allow multiple LOD updates before fps is calcd
As LOD based updating of terrain is what kills time, so allow multiple
of those events to occur before fps is calculated and shown.

Now the fps should be bit more accurate, compared to the last commit,
which messed up fps calc, by following a new flow for updating things
in the code.