Update terrain only after traversing a given distance
This is more of what it should be. Currently it is set such that
as one travels roughly 10% of the terrain dimensions (ie widht and
or height), a update request is sent to the terrain.

NOTE: When the terrain is updating inturn it takes few seconds,
which is ok for now, later need to look at how to optimise things,
once the LOD blackhole that I am noticing is fixed.

However two things needing exploring are

The GeoMipTerrain LOD logic I am still not able to get a hold on.
With full bruteforce things are fine, but without it the near and
far lod distance doesnt seem to be working. Changing it doesnt seem
to impact things in any way, and everything appears rendered at a
very low quality wrt polygons.

Setting the x and y scale of terrain root, doesnt seem to increase
the terrain space (ie compared to others), in that when the camera
is at the boundaries of the terrain, its position corresponds to
the original size of the terrain, itself.