Reduced BlockSize and lodFar, Avoid XY terrain scaling
Rather how the terrain appears seems to be impacted by

a) Scaling of terrain if any (Even thou as noted b4, it doesnt
seem to change the dimensions wrt other entities like camera etc)

b) the blocksize (Smaller blocksize seems to make it more Finer
and pronounced in a sense)

Rather initially when I reduced the blocksize from 1k to 32, the
1st time the terrain is rendered/generated it seems to show the
distant things fuzzied but in a way fine. However once the update
is called, it seems to make things very flat.

Then noticied that without scaling, if I manintain the original
4k lodFar, the far away mountains become more pronounced. However
the calling of update flattening things still remains.

	So now I have removed the 4x scaling along x and y axis
	and inturn reduced the lodFar by a factor of 4.

Next need to check that AutoFlatten thing once. (Rather on avoiding
autoflatten, the terrain retains its finer details even after update,
however have reverted the change for this commit, to go about it
in steps, in the commit, and also bcas the mode of autoflatten
doesnt seem to be having a impact, I seem to need to disable flat
tening fully. But ideally I want neither full flatten nor No flatten
but something inbetween).