After breakingHead, hexDumpingGeotiffs and Reading Spec -SY
I had assumed that the ModelPixelScaleTag will explicitly specify
the -ve sign wrt SY, as by default the raster space and model space
move in opposite direction wrt vertical axis, normally and or when
ever that is the case.

However on interpreting a geotiff file by looking at TAGs returned
by pil and libtiff, I noticed that the SY for the default case of
mapping btw raster and model space (ie moving in oppositive dirs)
contains a positive value and not negative value.

As initially I had used a old version of pillow and also libtiff in
debian, so just to be sure also dumped the tiff file using xxd and
looked at the file offset returned by libtiff/pil wrt the tag and
inturn manually interpreted the Double Precision 64bit floating point
binary value (rather indpendent of it also, just looking at the 64th
bit (ie the sign bit) also confirms same) also indicated that the SY
is stored as +ve value, when one actually requires to interpret as a
-ve value. Rather even both SX and SY had same value when SY needed
to be -ve.

So the conclusion is that one needs to explicitly make the SY opposite
of what is given. So if +ve make it -ve and if -ve make it +ve.

So have updated the logic in imgutils to do this negation explicitly.

Even reading the co-ord transform part of OSC GeoTiff spec, seems to
indicate the need to negate.