FSim: Terrain scaleZ only 100x; Proper Height within terrain
Scale only to 100x instead of 200x along z(height) axis.

Account for this scaling when calculating the height of terrain at
any given position.

Display height only when within the terrain. Else set a large -ve
value which wont occur in normal operation.

NOTE: the heightfield map could have a value of 0 to 1 or maybe
-1 to 1. And with the default 100x scaling it can become maximum
of -100 to 100.