ImgUtils: Init Config Dict globally
Now the default config values are initialised globally by default.
So even if one doesnt call handle_args or init to set or change
the config params used by the module, it will still work, with
the default config.

Also if no explicit cfg dict is passed to handle_args, then it
automatically uses the modules global cfg dict. So no need for
seperate boolean to tell whether to init the module's gCfg or not.

Now init no longer replaces the global cfg dict, rather it adds to
it.