FSim: Trying using thread for Objects hide and seek
As I am not adding or removing objects into the scene graph, but
only changing property of already added/existing nodepaths, so I
assume this should be fine at some levels.

Not sure if GeoMipTerrain update can also be moved into the thread
or not, as the documentation is not clear about it. Also there was
some discussion in the forum that it may not be threadsafe.