FSim: Test terrain update from a thread
Even thou moved to a different thread from main thread, the flow
still seems to freeze when the terrain update is triggered, most
probably because internally the screen draw and update logics of
terrain are handshaking through a lock and inturn working on a
single copy of the terrain rather than update working ona duplicate
and then switching over once done.

This is especially pronounced when flatten is set on the terrain.
Either way as flatten seems to fully flatten the terrain, so the
default AFMode is AFMOff.